[FEA] add CRUD functions
