+++ CareerPathLog/Model/JobOffer.swift
@@ -6,10 +6,16 @@ struct JobOffer: Codable {
     let jobTitle: String
     let urlOffer: String?
     let notes: String?
+    var remote: Bool
     let dateOfSentCV: Date
+    var reply: Bool
     let dateOfReply: Date?
-    let firstRoundOfInterview: Date?
-    let secondRoundOfInterview: Date?
+    var firstRoundInterview: Bool
+    let dateOfFirstRoundInterview: Date?
+    var secondRoundInterview: Bool
+    let dateOfSecondRoundInterview: Date?
+    var thirdRoundInterview: Bool
+    let dateOfThirdRoundInterview: Date?
     let fullTextOffer: String?
     var jobOfferRejected = false
 }
